winit: handle window events in WinitWindowAdapter
EventLoopState::window_event now only resolves the window and runs the
CustomApplicationHandler hook; the per-window work moved into
WinitWindowAdapter::dispatch_winit_window_event, which returns its error
instead of reaching up to set the event loop's error field.

The input state it needs (last cursor position, whether a mouse button is
down, the drag-resize direction, and the iOS touch id allocator) lived on
EventLoopState and was therefore shared by every window. It's now per
window, which fixes a two possible multi-window bugs:

- wheel, pinch and rotation events used the cursor position last seen in
  whichever window moved most recently, not their own. Not very likely
  to happen, but for example if a window suddenly pops up during
  "wheeling", then there might not be a move.
- a mouse button held in one window suppressed the pointer-exit event of
  another window the pointer left. Maaaybe if a dialog closes on press,
  something is highlighted in the window underneath, cursor moved out
  but the exit event isn't delivered because pressed as still true.

The coalesced mouse move stays in a single slot, now in SharedBackendData
so the adapter can reach it. Buffering a move for a different window used
to overwrite and drop the pending one; it's dispatched first instead.
